+++ gantry/commands/__init__.py
@@ -1,4 +1,5 @@
 from .cluster import cluster
+from .completion import completion
 from .config import config
 from .follow import follow
 from .list import list_cmd
@@ -6,4 +7,4 @@ from .main import main
 from .run import run
 from .stop import stop
 
-__all__ = ["main", "follow", "run", "config", "cluster", "stop", "list_cmd"]
+__all__ = ["main", "follow", "run", "config", "cluster", "stop", "list_cmd", "completion"]