+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `gantry completion ...` commands for configuring `gantry` shell autocompletion.
+
 ## [v1.3.0](https://github.com/allenai/beaker-gantry/releases/tag/v1.3.0) - 2024-05-31
 
 ### Added